[FEAT] 별점 step 수정

+++ app/src/main/java/com/example/Affiliates/data/CreateReview.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class CreateReview(
     @SerializedName(value = "review") val review: String,
-    @SerializedName(value = "star") val star: Float,
+    @SerializedName(value = "star") val star: Int,
     @SerializedName(value = "storeIdx") val storeIdx: Int
 )
